inventra-main/inventra-main/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.1
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.26.0
python-dotenv==1.0.0
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
python-multipart==0.0.6